# lib/swagger/doc/internal.yaml
swagger: '2.0'
info:
  description: Helo Whatsapp API documentation
  version: 1.0.0
  title: Helo Whatsapp
  termsOfService: ''
  contact:
    name: vivaconnect support
    email: support@vivaconnect.co
...
schemes:
  - http
paths:
...
  /users/account/tokenKey:
    put:
      description: API is used for updating user account token key
      summary: update user account token key
      tags:
        - accountProfile
      operationId: generateAndUpdateTokenKey
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: false
          description: ''
          schema:
            $ref: '#/definitions/updateAccountTokenKeyRequest'
      responses:
        '200':
          description: Returns token key generated
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 2000
              msg: Success
              data:
                tokenKey: sdvabsfbsbfdsdf3424fe342
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
definitions:
...
  updateAccountTokenKeyRequest:
    title: updateAccountInfoRequest
    example: {}
    type: object
    properties: {}
...